+++ ts/packages/providers/vercel/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @composio/vercel
 
+## 0.1.34
+
+### Patch Changes
+
+- Add exports for connection request
+- Updated dependencies
+  - @composio/core@0.1.34
+
+## 0.1.33
+
+### Patch Changes
+
+- Fix types and exports
+- Updated dependencies
+  - @composio/core@0.1.33
+
+## 0.1.32
+
+### Patch Changes
+
+- Add support for File type in uploading files
+- Updated dependencies
+  - @composio/core@0.1.32
+
 ## 0.1.31
 
 ### Patch Changes